# api/swaggerDoc.json
{
  "swagger": "2.0",
  "info": {
    "title": "API Gestor de Frotas",
    "description": "API para controle de frotas de veículos",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "tags": [
    {
      "name": "Veículos",
      "description": "Controle de Veículos"
    },
    {
      "name": "Setores",
      "description": "Controle de Setores"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/veiculos/": {
      "get": {
        "tags": [
          "Veículos"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Veículos"
        ],
        "description": "",
        "parameters": [
          {
            "name": "Veiculo",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "placa": {
                  "type": "string",
                  "example": "Placa do veículo"
                },
                "marca": {
                  "type": "string",
                  "example": "Marca do veículo"
                },
                "modelo": {
                  "type": "string",
                  "example": "Modelo do veículo"
                },
                "tipo": {
                  "type": "string",
                  "example": "Tipo do veículo"
                },
                "status": {
                  "type": "string",
                  "example": "Status do veículo"
                },
                "chassi": {
                  "type": "string",
                  "example": "Chassi do veículo"
                },
                "renavam": {
                  "type": "string",
                  "example": "Renavam do veículo"
                },
                "km": {
                  "type": "number",
                  "example": 0
                },
                "ano": {
                  "type": "number",
                  "example": 0
                },
                "cor": {
                  "type": "string",
                  "example": "Cor do veículo"
                },
                "combustivel": {
                  "type": "string",
                  "example": "Combustível do veículo"
                },
                "categoriaCNH": {
                  "type": "string",
                  "example": "Categoria de CNH para dirigir o veículo"
                },
                "dataLicenciamento": {
                  "type": "string",
                  "example": "Data do ultimo licenciamento do veículo"
                },
                "numPatrimonio": {
                  "type": "number",
                  "example": 0
                },
                "setorId": {
                  "type": "number",
                  "example": 0
                }
              },
              "required": [
                "placa",
                "marca",
                "modelo",
                "tipo",
                "status",
                "chassi",
                "renavam",
                "km",
                "ano",
                "cor",
                "combustivel",
                "categoriaCNH",
                "dataLicenciamento",
                "numPatrimonio",
                "setorId"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/veiculos/{id}": {
      "get": {
        "tags": [
          "Veículos"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "Veículos"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "Veiculo",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "placa": {
                  "type": "string",
                  "example": "Placa do veículo"
                },
                "marca": {
                  "type": "string",
                  "example": "Marca do veículo"
                },
                "modelo": {
                  "type": "string",
                  "example": "Modelo do veículo"
                },
                "tipo": {
                  "type": "string",
                  "example": "Tipo do veículo"
                },
                "status": {
                  "type": "string",
                  "example": "Status do veículo"
                },
                "chassi": {
                  "type": "string",
                  "example": "Chassi do veículo"
                },
                "renavam": {
                  "type": "string",
                  "example": "Renavam do veículo"
                },
                "km": {
                  "type": "number",
                  "example": 0
                },
                "ano": {
                  "type": "number",
                  "example": 0
                },
                "cor": {
                  "type": "string",
                  "example": "Cor do veículo"
                },
                "combustivel": {
                  "type": "string",
                  "example": "Combustível do veículo"
                },
                "categoriaCNH": {
                  "type": "string",
                  "example": "Categoria de CNH para dirigir o veículo"
                },
                "dataLicenciamento": {
                  "type": "string",
                  "example": "Data do ultimo licenciamento do veículo"
                },
                "numPatrimonio": {
                  "type": "number",
                  "example": 0
                },
                "setorId": {
                  "type": "number",
                  "example": 0
                }
              },
              "required": [
                "placa",
                "marca",
                "modelo",
                "tipo",
                "status",
                "chassi",
                "renavam",
                "km",
                "ano",
                "cor",
                "combustivel",
                "categoriaCNH",
                "dataLicenciamento",
                "numPatrimonio",
                "setorId"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Veículos"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/setores/": {
      "get": {
        "tags": [
          "Setores"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Setores"
        ],
        "description": "",
        "parameters": [
          {
            "name": "Setor",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "type": "string",
                  "example": "Nome do setor"
                }
              },
              "required": [
                "nome"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/setores/{id}": {
      "get": {
        "tags": [
          "Setores"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "Setores"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "Setor",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "type": "string",
                  "example": "Nome do setor"
                }
              },
              "required": [
                "nome"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Setores"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  }
}